+++ language-ats/src/Language/ATS/Rewrite.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE FlexibleContexts #-}
+
 module Language.ATS.Rewrite ( rewriteDecl
                             -- * Fixity
                             , defaultFixityState
@@ -13,7 +15,10 @@ import           Language.ATS.Types.Lens
 import           Lens.Micro
 
 exprLens :: Eq a => FixityState a -> ASetter s t (Expression a) (Expression a) -> s -> t
-exprLens st = flip over (rewriteATS st)
+exprLens = flip over . rewriteATS
+
+exprStaLens :: Eq a => FixityState a -> ASetter s t (StaticExpression a) (StaticExpression a) -> s -> t
+exprStaLens = flip over . rewriteStaATS
 
 exprLenses :: Eq a => FixityState a -> [ASetter b b (Expression a) (Expression a)] -> b -> b
 exprLenses st = thread . fmap (exprLens st)
@@ -22,6 +27,8 @@ rewriteDecl :: Eq a => FixityState a -> Declaration a -> Declaration a
 rewriteDecl st (Extern l d) = Extern l (rewriteDecl st d)
 rewriteDecl st x@Val{} = exprLens st valExpression x
 rewriteDecl st x@Var{} = exprLenses st [varExpr1._Just, varExpr2._Just] x
+rewriteDecl st x@PrVal{} = exprStaLens st (prValExpr._Just) x
+rewriteDecl st x@PrVar{} = exprStaLens st (prValExpr._Just) x
 rewriteDecl st x@Func{} = exprLens st (fun.preF.expression._Just) x
 rewriteDecl st x@Impl{} = exprLens st (impl.iExpression._Right) x
 rewriteDecl st x@AndDecl{} = exprLens st andExpr x
@@ -77,6 +84,14 @@ getFixity st (SpecialInfix _ op') =
 -- FIXME this should account for right vs. left associativity.
 compareFixity :: Eq a => FixityState a -> BinOp a -> BinOp a -> Bool
 compareFixity st = (== GT) .* on compare (getFixity st)
+
+rewriteStaATS :: Eq a => FixityState a -> StaticExpression a -> StaticExpression a
+rewriteStaATS st = cata a where
+    a (SCallF n ts [StaticVoid{}]) = SCall n ts []
+    a (StaticBinaryF op (StaticBinary op' e e') e'')
+        | compareFixity st op op'  = StaticBinary op e (StaticBinary op' e' e'')
+    a (WhereStaExpF se (ATS ds))   = WhereStaExp se (ATS (rewriteDecl st <$> ds))
+    a x                            = embed x
 
 -- | Among other things, this rewrites expressions so that operator precedence
 -- is respected; this ensures @1 + 2 * 3@ will be parsed as the correct